refactor: Modularization of the carousels, course and communities.
