+++ Pages/HomePage/style.css
@@ -303,14 +303,14 @@ header .Navbar {
     background-color: var(--blue-hover-bright);
 }
 
-.Cursos-Carousel{
+.Carousel{
     position: relative;
     width: 100%;
     max-width: 2000px;
     overflow: hidden;
 }
 
-.CursosItems {
+.CourseList {
     display: flex;
     flex-direction: row;
     flex-wrap: nowrap;
@@ -341,27 +341,34 @@ header .Navbar {
     right: 0;
 }
 
-.CursoItem {
-
+.CourseItem {
     margin: 20px 0;
+    padding: 10px;
+    border-radius: 20px;
     min-width: 240px;
     max-width: 270px;
     width: 20%;
     cursor: pointer;
+    transition: 0.2s ease;
 }
 
-.CursoItem .CursoImage {
+.CourseItem:hover{
+    box-shadow: 0px 30px 80px rgba(0, 0, 0, 0.2);
+}
+
+.CourseItem .CursoImage {
+    border-radius: 10px;
     max-width: 270px;
     width: 100%;
 }
 
-.CursoItem .CursoTitle {
+.CourseItem .CursoTitle {
     font-size: 16px;
     line-height: 20px;
     margin-top: 10px;
 }
 
-.CursoItem .CursoAuthor {
+.CourseItem .CursoAuthor {
     color: var(--gray-hover-dark);
     font-size: 16px;
     line-height: 20px;
@@ -397,13 +404,7 @@ header .Navbar {
     color: var(--black);
 }
 
-.CommunityCarousel{
-    position: relative;
-    width: 100%;
-    overflow: hidden;
-}
-
-.CommunityItems{
+.CommunityList{
     width: 100%;
     margin-top: 40px;
     display: flex;
@@ -415,6 +416,7 @@ header .Navbar {
 }
 
 .CommunityItem {
+    margin: 20px;
     display: flex;
     flex-direction: column;
     align-items: center;
@@ -428,6 +430,10 @@ header .Navbar {
     text-align: center;
     color: white;
     margin-top: 10px;
+}
+
+.CommunityItem:hover{
+    box-shadow: 0px 30px 80px rgba(0, 0, 0, 0.2);
 }
 
 .CommunityImage {
